refactor(#3267): extract builtin static/prototype VALUE-read subsystem from property-access.ts
Pure god-file split (subtask of #3182). Moves the cohesive built-in
static/prototype VALUE-read group — the Math/Number/TypedArray/ctor metadata
tables plus the #1907/#1888-S6-b standalone value-read machinery — verbatim into
a new sibling module src/codegen/builtin-value-read.ts. No logic changes.

property-access.ts: 8937 -> 7990 LOC. New module: 1058 LOC (19 symbols).
The new module imports only leaf helpers (zero back-edge into property-access.ts,
no import cycle); property-access.ts imports back the 12 symbols it still calls
and re-exports the 9 that calls.ts / builtin-static-gopd.ts import from it.

Behaviour gate: prove-emit-identity check -> IDENTICAL (39/39 emits across
gc/standalone/wasi). tsc --noEmit -> 0. tests/issue-3267.test.ts (7 standalone
smoke tests) green. Relocation-shift: 2 relocated checker sites preauthorized in
oracle-ratchet-baseline.json (additive, per the #808 precedent); loc-budget,
coercion-sites, dead-exports, verdict-oracle-bump all pass unchanged.


✓
